requirements.txt
flask>=2.3.0
requests>=2.31.0
pywebview>=4.0.0
